bgp_direct_to_bgp_vpn: convert to lib/ltemplate
